+++ apps/ontology-explorer/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @party-stack/ontology-explorer
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [46268bc]
+    - @party-stack/ontology@0.10.0
+    - @party-stack/foundry-ontology@0.9.0
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [bc97879]
+    - @party-stack/foundry-ontology@0.8.1
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [515f8dc]
+    - @party-stack/ontology@0.9.0
+    - @party-stack/foundry-ontology@0.8.0
+
 ## 0.0.10
 
 ### Patch Changes