+++ packages/ontology-devtools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @party-stack/ontology-devtools
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [46268bc]
+    - @party-stack/ontology@0.10.0
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [515f8dc]
+    - @party-stack/ontology@0.9.0
+
 ## 0.3.0
 
 ### Minor Changes